adding small headers to the socket packets
JSSocketModule now prepends the length of the JSON string and the crc32 hash so
we can then verify that what we've received is complete on the other end. NodeJS
side checks for these values, and - for now - if the JSON string received is
long enough it is parsed. If it's not long enough wait for the rest of the data
before parsing. We'll have to handle a timeout for incomplete packets.